reorder pre-commit hooks to run blue before flake8
