+++ .pre-commit-config.yaml
@@ -9,16 +9,16 @@ repos:
     -  id: end-of-file-fixer
     -  id: trailing-whitespace
 
+  - repo: https://github.com/grantjenks/blue
+    rev: v0.9.1
+    hooks:
+      -  id: blue
+
   - repo: https://github.com/PyCQA/flake8
     rev: 5.0.4
     hooks:
       -  id: flake8
          language_version: python3
-
-  - repo: https://github.com/grantjenks/blue
-    rev: v0.9.1
-    hooks:
-      -  id: blue
 
   - repo: https://github.com/asottile/reorder_python_imports
     rev: v3.8.2
